requirements.txt
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
autodistill==0.1.24
autodistill-fastsam==0.1.1
autodistill-grounded-sam==0.1.2
Automat==22.10.0
beautifulsoup4==4.12.2
captcha==0.5.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
constantly==23.10.4
contourpy==1.2.0
cryptography==41.0.7
cycler==0.10.0
daphne==4.0.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.9
django-allauth==0.58.2
django-bootstrap-v5==1.0.11
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.109.0
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.44.0
frozenlist==1.4.1
fsspec==2023.10.0
gradio==3.35.2
gradio_client==0.8.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
hyperlink==21.0.0
idna==2.10
importlib-metadata==7.0.1
incremental==22.10.0
ipykernel==6.29.0
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
labelImg==1.8.6
linkify-it-py==2.0.2
lxml==4.9.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
nest-asyncio==1.5.9
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
onnx==1.15.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.0.74
orjson==3.9.12
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
pipenv==2023.10.24
platformdirs==3.11.0
ply==3.11
prompt-toolkit==3.0.43
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.11
PyQt5-sip==12.13.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rf-segment-anything==1.0
rf_groundingdino==0.1.2
roboflow==1.1.16
rpds-py==0.17.1
safetensors==0.4.1
scikit-learn==1.4.0
scipy==1.11.3
seaborn==0.13.0
semantic-version==2.10.0
service-identity==24.1.0
sip==6.7.12
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.35.1
supervision==0.9.0
sympy==1.12
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
timm==0.9.12
tokenizers==0.15.0
tomli==2.0.1
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
ultralytics==8.1.3
urllib3==2.1.0
uvicorn==0.26.0
virtualenv==20.24.6
wcwidth==0.2.13
websockets==11.0.3
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
zope.interface==6.1
